chore(client): remove unused deps and rename files to kebab-case
