+++ apps/client/package.json
@@ -19,10 +19,7 @@
     "react": "^19.2.5",
     "react-dom": "^19.2.5",
     "react-router": "^7.14.2",
-    "react-router-dom": "^7.14.2",
-    "@react-router/node": "^7.14.2",
-    "zustand": "^5.0.12",
-    "isbot": "^5"
+    "zustand": "^5.0.12"
   },
   "devDependencies": {
     "@babel/core": "^7.29.0",
@@ -41,7 +38,6 @@
     "@vitejs/plugin-react": "^6.0.1",
     "@vitest/coverage-v8": "^4.1.5",
     "babel-plugin-react-compiler": "^1.0.0",
-    "globals": "^17.5.0",
     "jsdom": "^29.0.2",
     "typescript": "~6.0.2",
     "vite": "^8.0.9",
