+++ apps/client/tests/common/router.test.tsx
@@ -4,7 +4,7 @@ import { cleanup, render, screen } from "@testing-library/react";
 import { createMemoryRouter, RouterProvider } from "react-router";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { useUserAuthStore } from "#/features/auth/store/userAuthStore";
+import { useUserAuthStore } from "#/features/auth/store/user-auth-store";
 import AppLayout from "#/routes/_app";
 import IndexRoute from "#/routes/_index";
 import LobbyRoute from "#/routes/lobby";
